// packages/frontend/src/openapi/dnd-api/model/gameResponseDto.ts
import type { Board } from "./board";
import type { GameMasterResponseDto } from "./gameMasterResponseDto";
import type { GameResponseDtoPlayableEntitiesItem } from "./gameResponseDtoPlayableEntitiesItem";
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */
import type { GameStatus } from "./gameStatus";

export interface GameResponseDto {
  status: GameStatus;
  playableEntities: GameResponseDtoPlayableEntitiesItem[];
  id: string;
  board: Board;
  gameMaster: GameMasterResponseDto;
  timeline: string[];
}