// packages/frontend/src/openapi/dnd-api/model/interactiveEntityKind.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */

export type InteractiveEntityKind =
  (typeof InteractiveEntityKind)[keyof typeof InteractiveEntityKind];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InteractiveEntityKind = {
  door: "door",
  trap: "trap",
  chest: "chest",
} as const;